requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
gunicorn==21.2.0
flask-limiter==3.5.0
marshmallow==3.20.1
redis==4.6.0
celery==5.3.6
flask-socketio==5.3.6
transformers==4.46.0
torch==2.1.2
huggingface==0.26.1
sentence-transformers==2.11.0
accelerate==0.25.0
